backend/foodgram/users/views.py
from rest_framework import viewsets
from rest_framework.pagination import LimitOffsetPagination

from .models import User
from .serializers import UserSerializer


class UserViewSet(viewsets.ModelViewSet):
    """Users' viewset."""
    queryset = User.objects.all()
    serializer_class = UserSerializer
    pagination_class = LimitOffsetPagination
